issue fix

+++ administrator/models.py
@@ -192,7 +192,7 @@ class CsvLogDetails(models.Model):
     age = models.IntegerField(null=True, blank=True)
     system_name = models.CharField(null=True, blank=True, max_length=1000)
     factors_effected = models.CharField(null=True, blank=True, max_length=1000)
-    downtime_week = models.IntegerField(default=0, null=True, blank=True)
+    downtime_week = models.FloatField(default=0, null=True, blank=True)
     impact_hour = models.CharField(null=True, blank=True, max_length=1000)
 
     hourly_rate = models.FloatField(null=True, blank=True, default=0)
